driver: address some warnings warnings on 64-bit
Problem: compiler emits warnings about cast to integer of different
size, etc. when building on x86_64.

Fix up some types, all in the construction of printk parameters,
so should have little effect.

+++ driver/module.c
@@ -521,7 +521,7 @@ loff_t *off;
   blocking = (dev->Uart.block & SI_SERIAL_FLAGS_BLOCK)!=0;
 
   if( dev->verbose & SI_VERBOSE_SERIAL)
-    printk("SI write, count %d\n", count );
+    printk("SI write, count %lu\n", (unsigned long)count);
 
   if( dev->test ) {
     for (i=0; i < count; i++) {     // for all characters
